Remove TrackType.double, complete migration to paired singles
- Remove `case double` from Edge.TrackType enum (3 cases remain:
  highSpeed, regional, single)
- Add legacy decoder in TrackType.init(from:): JSON "double" →
  .single so old files load without crashing; migrateDoubleTracksToSingle
  treats them as undirected single tracks (correct pathfinding)
- Update Edge+Migration: guard now checks only .highSpeed
- Fix all switch exhaustiveness: 22 files updated
  (UI editors, renderers, inspectors, domain services)
- Replace .double capacity logic in RailwayGraphManager with
  pairedEdgeId != nil check (canonical post-migration signal)
- Remove .double buttons from all track-type pickers (5 views)
- SchedulingServicesTests: .double → .single fixtures
- EdgeMigrationTests: repurposed case 1 (multi-highSpeed),
  cases 3/5 use .highSpeed as migration trigger

+++ FdC Railway Manager/Infrastructure/IO/IOManager.swift
@@ -97,7 +97,7 @@ final class IOManager: ObservableObject {
                 switch fdc.trackType?.lowercased() {
                 case "highspeed", "high_speed": return .highSpeed
                 case "single": return .single
-                case "double": return .double
+                case "double": return .single  // legacy — migrated to paired singles at load
                 default: return .regional
                 }
             }()
